requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
fastapi==0.120.0
filelock==3.20.0
fsspec==2025.10.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.2
langchain-groq==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
openai==2.7.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
portalocker==3.2.0
protobuf==6.33.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
